Move Query and Body Params to strategies
TODO: move the "clean" functionality
